requirements.txt.txt

colorama==0.4.6
prettytable==3.9.0
requests==2.31.0
scapy==2.5.0
cryptography==42.0.5
dnspython==2.6.1
whois==0.9.5
tqdm==4.66.4
beautifulsoup4==4.12.3
paramiko==3.4.0          
PyJWT==2.8.0            
googlesearch-python==1.2.3  
python-whois==0.9.5      
pyOpenSSL==24.0.0       
impacket==0.11.0
dnspython==2.6.1